feat: add app panel for customers
