+++ app/Filament/Admin/Resources/EventResource/Pages/CreateEvent.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Admin\Resources\EventResource\Pages;
+
+use App\Filament\Admin\Resources\EventResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateEvent extends CreateRecord
+{
+    protected static string $resource = EventResource::class;
+
+    protected function mutateFormDataBeforeCreate(array $data): array
+    {
+        $data['user_id'] = auth()->id();
+        $data['slug'] = str()->slug($data['title']);
+
+        return $data;
+    }
+}
