+++ app/Filament/Admin/Resources/MessageResource/Pages/ManageMessages.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Filament\Admin\Resources\MessageResource\Pages;
+
+use App\Filament\Admin\Resources\MessageResource;
+use Filament\Actions;
+use Filament\Resources\Pages\ManageRecords;
+
+class ManageMessages extends ManageRecords
+{
+    protected static string $resource = MessageResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\CreateAction::make()
+                ->modalWidth('xl'),
+        ];
+    }
+}
