--- app/Filament/Resources/EventResource/Pages/CreateEvent.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\EventResource\Pages;
-
-use App\Filament\Resources\EventResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateEvent extends CreateRecord
-{
-    protected static string $resource = EventResource::class;
-
-    protected function mutateFormDataBeforeCreate(array $data): array
-    {
-        $data['user_id'] = auth()->id();
-        $data['slug'] = str()->slug($data['title']);
-
-        return $data;
-    }
-}
