--- app/Filament/Resources/MessageResource.php
@@ -1,97 +1,0 @@
-<?php
-
-namespace App\Filament\Resources;
-
-use App\Enums\EventType;
-use App\Filament\Resources\MessageResource\Pages;
-use App\Models\Message;
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\SoftDeletingScope;
-
-class MessageResource extends Resource
-{
-    protected static ?string $model = Message::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
-
-    protected static ?int $navigationSort = 2;
-
-    public static function getModelLabel(): string
-    {
-        return trans_choice('ohyez.message', 1);
-    }
-
-    public static function getPluralModelLabel(): string
-    {
-        return trans_choice('ohyez.message', 2);
-    }
-
-    public static function getNavigationGroup(): ?string
-    {
-        return trans_choice('ohyez.content', 2);
-    }
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->columns(1)
-            ->schema([
-                Forms\Components\Select::make('event_type')
-                    ->options(EventType::class)
-                    ->required(),
-                Forms\Components\Textarea::make('content')
-                    ->rows(6)
-                    ->required(),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('event_type')
-                    ->badge()
-                    ->formatStateUsing(fn ($state) => __('ohyez.'.$state))
-                    ->sortable(),
-                Tables\Columns\TextColumn::make('content')
-                    ->wrap(),
-            ])
-            ->filters([
-                Tables\Filters\TrashedFilter::make(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make()
-                    ->modalWidth('xl'),
-                Tables\Actions\DeleteAction::make(),
-                Tables\Actions\ForceDeleteAction::make(),
-                Tables\Actions\RestoreAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                    Tables\Actions\ForceDeleteBulkAction::make(),
-                    Tables\Actions\RestoreBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ManageMessages::route('/'),
-        ];
-    }
-
-    public static function getEloquentQuery(): Builder
-    {
-        return parent::getEloquentQuery()
-            ->withoutGlobalScopes([
-                SoftDeletingScope::class,
-            ]);
-    }
-}
