--- app/Filament/Resources/MessageResource/Pages/ManageMessages.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\MessageResource\Pages;
-
-use App\Filament\Resources\MessageResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ManageRecords;
-
-class ManageMessages extends ManageRecords
-{
-    protected static string $resource = MessageResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make()
-                ->modalWidth('xl'),
-        ];
-    }
-}
