--- app/Filament/Resources/TemplateResource.php
@@ -1,96 +1,0 @@
-<?php
-
-namespace App\Filament\Resources;
-
-use App\Filament\Resources\TemplateResource\Pages;
-use App\Models\Template;
-use App\Settings\GeneralSettings;
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\SoftDeletingScope;
-
-class TemplateResource extends Resource
-{
-    protected static ?string $model = Template::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
-
-    public static function getModelLabel(): string
-    {
-        return trans_choice('ohyez.template', 1);
-    }
-
-    public static function getPluralModelLabel(): string
-    {
-        return trans_choice('ohyez.template', 2);
-    }
-
-    public static function getNavigationGroup(): ?string
-    {
-        return trans_choice('ohyez.content', 2);
-    }
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->columns(1)
-            ->schema([
-                Forms\Components\TextInput::make('name')
-                    ->required(),
-                Forms\Components\Select::make('event_type')
-                    ->options(collect(app(GeneralSettings::class)->event_types)
-                        ->mapWithKeys(fn ($item) => [$item => __('ohyez.'.$item)])
-                    )
-                    ->required(),
-                Forms\Components\TextInput::make('view')
-                    ->default('ohyez::templates.default')
-                    ->required(),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('event_type')
-                    ->formatStateUsing(fn ($state) => __('ohyez.'.$state))
-                    ->badge(),
-                Tables\Columns\TextColumn::make('name'),
-            ])
-            ->filters([
-                Tables\Filters\TrashedFilter::make(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make()->modalWidth('md'),
-                Tables\Actions\DeleteAction::make(),
-                Tables\Actions\ForceDeleteAction::make(),
-                Tables\Actions\RestoreAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                    Tables\Actions\ForceDeleteBulkAction::make(),
-                    Tables\Actions\RestoreBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ManageTemplates::route('/'),
-        ];
-    }
-
-    public static function getEloquentQuery(): Builder
-    {
-        return parent::getEloquentQuery()
-            ->withoutGlobalScopes([
-                SoftDeletingScope::class,
-            ]);
-    }
-}
