--- app/Filament/Resources/UserResource/Pages/CreateUser.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\UserResource\Pages;
-
-use App\Filament\Resources\UserResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateUser extends CreateRecord
-{
-    protected static string $resource = UserResource::class;
-}
