--- app/Filament/Resources/UserResource/Pages/EditUser.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\UserResource\Pages;
-
-use App\Filament\Resources\UserResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditUser extends EditRecord
-{
-    protected static string $resource = UserResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\DeleteAction::make(),
-            Actions\ForceDeleteAction::make(),
-            Actions\RestoreAction::make(),
-        ];
-    }
-}
