+++ bootstrap/providers.php
@@ -2,6 +2,7 @@
 
 return [
     App\Providers\AppServiceProvider::class,
-    App\Providers\Filament\AdminPanelProvider::class,
     App\Providers\AuditFieldsSchemaServiceProvider::class,
+    App\Providers\Filament\AdminPanelProvider::class,
+    App\Providers\Filament\AppPanelProvider::class,
 ];
