+++ resources/css/filament/app/tailwind.config.js
@@ -3,9 +3,8 @@ import preset from '../../../../vendor/filament/filament/tailwind.config.preset'
 export default {
     presets: [preset],
     content: [
-        './app/Filament/admin/**/*.php',
-        './resources/views/filament/admin/**/*.blade.php',
-        './resources/views/livewire/**/*.blade.php',
+        './app/Filament/App/**/*.php',
+        './resources/views/filament/app/**/*.blade.php',
         './vendor/filament/**/*.blade.php',
     ],
 }
