+++ resources/views/filament/admin/resources/event/pages/attendance.blade.php
