+++ resources/views/filament/admin/resources/event/pages/invitations.blade.php
