+++ resources/views/filament/admin/resources/event/pages/list-events.blade.php
