+++ resources/views/filament/app/pages/dashboard.blade.php
@@ -1,0 +1,7 @@
+<x-filament-panels::page>
+    <div>
+        <h1 class="text-xl">
+            TODO: implement dashboard
+        </h1>
+    </div>
+</x-filament-panels::page>
