+++ vite.config.js
@@ -4,7 +4,12 @@ import laravel, {refreshPaths} from 'laravel-vite-plugin';
 export default defineConfig({
     plugins: [
         laravel({
-            input: ['resources/css/app.css', 'resources/js/app.js', 'resources/css/filament/admin/theme.css'],
+            input: [
+                'resources/css/app.css',
+                'resources/js/app.js',
+                'resources/css/filament/admin/theme.css',
+                'resources/css/filament/app/theme.css'
+            ],
             refresh: [
                 ...refreshPaths,
                 'app/Livewire/**',
